+++ src/crypt/options.ts
@@ -3,7 +3,7 @@ import { CRYPT_DEFAULTS } from './defaults.js';
 
 /** Options for the crypt API. */
 export interface CryptOptions extends Omit<KeyDerivationOptions, 'instance'> {
-  /** Nonce or initialization vector. */
+  /** The nonce A.K.A. initialization vector (IV). */
   nonce: Uint8Array<ArrayBuffer>;
 }
 
@@ -11,7 +11,7 @@ export interface CryptOptions extends Omit<KeyDerivationOptions, 'instance'> {
  * Builds a full options object from the given options partial. Populates omitted properties with
  * defaults.
  *
- * If any of `encAlg`, `hashAlg`, `iterations`, `kdf`, or `keyLen` are not provided, its value from
+ * If any of `encAlg`, `hashAlg`, `iterations`, `keyLen`, or `kdf` are not provided, its value from
  * {@link CRYPT_DEFAULTS} will be used.
  *
  * If `nonce` is not provided, 12 random bytes will be generated.
@@ -24,14 +24,9 @@ export interface CryptOptions extends Omit<KeyDerivationOptions, 'instance'> {
  * @returns A full {@link CryptOptions} object.
  */
 export const cryptOptions = (partialOptions: Partial<CryptOptions>): CryptOptions => ({
-  encAlg: partialOptions.encAlg ?? CRYPT_DEFAULTS.encAlg,
-  hashAlg: partialOptions.hashAlg ?? CRYPT_DEFAULTS.hashAlg,
-  iterations: partialOptions.iterations ?? CRYPT_DEFAULTS.iterations,
-  kdf: partialOptions.kdf ?? CRYPT_DEFAULTS.kdf,
-  keyLen: partialOptions.keyLen ?? CRYPT_DEFAULTS.keyLen,
+  ...CRYPT_DEFAULTS,
+  ...partialOptions,
   nonce: partialOptions.nonce ?? crypto.getRandomValues(new Uint8Array(12)),
-  passphrase: partialOptions.passphrase,
-  publicKey: partialOptions.publicKey,
   salt: partialOptions.salt ?? crypto.getRandomValues(new Uint8Array(16)),
 });
 
@@ -40,8 +35,10 @@ export const cryptOptions = (partialOptions: Partial<CryptOptions>): CryptOption
  * @returns A sanitized version of the given {@link CryptOptions} with sensitive data removed. The
  *   original object remains unmodified.
  */
-export function sanitizeCryptOptions<T extends CryptOptions>(options: T): Omit<T, 'passphrase'> {
+export function sanitizeCryptOptions<T extends CryptOptions>(
+  options: T,
+): Omit<T, 'key' | 'passphrase'> {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { passphrase, ...sanitized } = options;
+  const { key, passphrase, ...sanitized } = options;
   return sanitized;
 }